Restore jedi.el in autoload
This module is used by elpy to `M-.` (go to definition) and stuffs.

I got confused and I removed it, but it's necessary.

On the other hand, I'm using `company-mode` for text completion and
`jedi` uses `auto-complete` (already removed) because of conflicts in
suggestions.

The problem is that `jedi` requires `auto-complete` and as I disable it
jedi is complaining, but still working without problems.