ERC reconnect if lost connection
I wrote this function and I expect to fail but I want to try it for a
while.

The idea is simple: check every 25 seconds if there is a buffer called
"#pyar" and if it has an ERC process running. If not, switch to it and
reconnect. Then switch back to the current buffer.